+++ src/data_readers.clj
@@ -1,0 +1,2 @@
+{datascript/Datom datascript/datom-from-reader
+ datascript/DB    datascript/db-from-reader}